Easier automation for students
When students book in the book.html, if they have only 1 student on their list, then automatically make them the primary student.

If student users add a student in their account, and the student that they are adding has the same first and last name as other students in their account, then automatically just update their previous student profile with the current one. So it keeps the same booked sessions, etc, and nothing will be deleted except for the student’s old info (pronouns, school, etc.)