home page of tutors is nice now
For tutors, add a "Booking Info" hyperlink button to the tutor's dashboard. Make the "booking-info.html" webpage info editable, so that tutors can edit their information (e.g. first student = $50, each additional student = $10)... Make it editable just like the Booking Info tab in the /admin.html

The tutor's page (/home/{tutorname}) should have all of the information of the tutor.
First it is the tutor's card (tutor-summary-card) which will include their headline and bio.

Then it will display the prices and availability (pretty much everything that is in the booking-info of that tutor).
Lastly, it will display the rest of their webpage that they can edit (all the editable widgets that is only editable by that tutor).